LayoutTests/http/tests/security/seamless/seamless-sandbox-srcdoc-expected.txt

FAIL iframe.seamless should be true (of type boolean). Was undefined (of type undefined).
PASS window.getComputedStyle(iframe).width is "200px"
PASS window.getComputedStyle(iframe).height is "100px"
